Cleaned up search header
